- patches.fixes/reiserfs-remove-2-tb-file-size-limit: Fix issue on   32-bit systems.
suse-commit: 7621175843901a97be5a91e05255ede5b66051bb

+++ fs/reiserfs/super.c
@@ -1313,7 +1313,7 @@ static inline loff_t
 reiserfs_max_file_offset(struct super_block *sb)
 {
 	/* Limited by stat_data->sd_blocks, 2^32-1 blocks */
-	loff_t fs_max = (sb->s_blocksize << 32) - sb->s_blocksize;
+	loff_t fs_max = ((u64)sb->s_blocksize << 32) - sb->s_blocksize;
 
 	/* Limited by 32-bit MAX_LFS_FILESIZE */
 	loff_t page_cache_max = (((u64)PAGE_CACHE_SIZE << 31)-1);
